+++ src/Game/StairsSceneryObject.h
@@ -1,12 +1,13 @@
 #pragma once
 
 #include "../Game/SceneryObject.h"
+#include "../Game/LadderSceneryObject.h"
 
 namespace Falltergeist
 {
     namespace Game
     {
-        class StairsSceneryObject : public SceneryObject
+        class StairsSceneryObject : public LadderSceneryObject
         {
             public:
                 StairsSceneryObject();