+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Under development
 - [feature] Extract UI generation logic from game objects (alexeevdv)
 - [feature] New video settings: window position(x,y) and always_on_top flag (adamkewley)
 - [feature] Added missing logo in the Pipboy window and special date greetings (JanSimek)
+- [feature] Added ladders and stairs funtionality (667bdrm)
 
 0.3.1 (2018-01-14)
 =======================